Treat chunked encoding as an error
As chunked encoding is not implemented and hasn't been handled properly,
it should be treated as an error for now.

According to the source code of [CFHTTPServer.c](https://apple.co/2TfWW5T),
it treats chunked encoding as an error as well.

Downgrading to HTTP 1.0 disables chunked encoding but we need to keep
using HTTP 1.1, which defines chunked encoding, to send Server-Sent
Events.