Remove direct dependency to babel
Since apollo-client-devtools is distributed in ES5 and it internally
uses babel, ApolloDeveloperKit need to depend on babel as compile time
dependency although apollo-client-devtools is only used for integration
testing.

This commit force apollo-client-devtools compiled with tsc.
Though it is theoretically wrong for the module compiled with babel to be
compiled with tsc, actually it is safe because apollo-client-devtools uses
babel only to translate the latest JavaScript syntax for older browser.

So I decided to quit depending on babel to reduce dependencies.